+++ zk_toolbox/crates/zk_supervisor/README.md
@@ -5,7 +5,7 @@ This document contains the help content for the `zk_supervisor` command-line pro
 **Command Overview:**
 
 - [`zk_supervisor`↴](#zk_supervisor)
-- [`zk_supervisor protocol-version`↴](#zk_supervisor-protocol-version)
+- [`zk_supervisor prover-version`↴](#zk_supervisor-prover-version)
 - [`zk_supervisor database`↴](#zk_supervisor-database)
 - [`zk_supervisor database check-sqlx-data`↴](#zk_supervisor-database-check-sqlx-data)
 - [`zk_supervisor database drop`↴](#zk_supervisor-database-drop)
@@ -45,11 +45,11 @@ ZK Toolbox is a set of tools for working with zk stack.
 - `--chain <CHAIN>` — Chain to use
 - `--ignore-prerequisites` — Ignores prerequisites checks
 
-## `zk_supervisor protocol-version`
+## `zk_supervisor prover-version`
 
 Gets information about current protocol version of provers in `zksync-era` and snark wrapper hash.
 
-**Usage:** `zk_supervisor protocol-version`
+**Usage:** `zk_supervisor prover-version`
 
 ## `zk_supervisor database`
