+++ zk_toolbox/crates/zk_supervisor/src/commands/mod.rs
@@ -3,6 +3,6 @@ pub mod database;
 pub mod fmt;
 pub mod lint;
 pub(crate) mod lint_utils;
-pub mod protocol_version;
+pub mod prover_version;
 pub mod snapshot;
 pub mod test;
