+++ zk_toolbox/crates/zk_supervisor/src/commands/prover_version.rs
